Prepare support for dynamic config overrides
Extract some config parameters into a separate config class. In a
follow-up PR I'll introduce a config provider that subscribes to an
MQTT topic to receive config updates. With this, we can make dynamic
configuration updates. For instance, I'd like to be able to
reconfigure the maximum battery wattage depending on my current
Tibber electricity price. If the price is low, I'd like to prefer
power from the grid. If the price is high, I'd like to increase the
power taken from the battery.